+++ src/App.jsx
@@ -1,7 +1,7 @@
 // import {PostForm} from "./component/postForm/PostForm";
-import { GetThought } from "./component/GetMessage/GetThought";
-import { Header } from "./component/GetMessage/header";
-import { Footer } from "./component/Footer";
+import { GetThought } from "./component/getThought/GetThought";
+import { Header } from "./component/header/header";
+import { Footer } from "./component/footer/Footer";
 
 export const App = () => {
   return (
