--- src/component/GetMessage/header.jsx
@@ -1,7 +1,0 @@
-export const Header = () => {
-    return (
-        <div>
-            <header>My Happy Thought</header>
-        </div>
-    )
-}
