+++ src/component/footer/Footer.jsx
@@ -1,3 +1,5 @@
+import './footer.css';
+
 export const Footer = () => {
     return (
         <div className="footer">
