+++ src/component/footer/footer.css
@@ -1,0 +1,4 @@
+footer{
+    text-align: center;
+    background-color: #F2F0F0;
+}
