+++ src/component/getThought/getThought.css
@@ -1,0 +1,36 @@
+
+.old-thought-container{
+    border: 1px solid black;
+    background-color: #FFFFFF;
+    box-shadow: 10px 10px black;
+    display: flex;
+    flex-direction: column;
+    flex-wrap: wrap;
+    align-items: baseline;
+    padding: 1rem; 
+    margin: 1rem; 
+    gap: 30px;
+}
+
+.heart-button {
+    background-color: #EAEAEA;
+    width: 2rem;
+    height:2rem;
+    border-radius: 50%;
+    cursor: pointer;
+    font-size: 1rem;
+    border: none;
+    margin-right: 0.5rem;
+    margin-bottom: 1rem;
+}
+
+.heart-button:active{
+    background-color: #FFADAD ;
+}
+
+.like-post-time {
+    color: #CACACA;
+}
+
+
+
