+++ src/component/header/header.css
@@ -1,0 +1,7 @@
+header {
+    font-size: 25px;
+    font-weight: 500;
+    height:50px;
+    margin-top: 20px;
+    text-align: center;
+}
