+++ src/component/header/header.jsx
@@ -1,0 +1,9 @@
+import './header.css';
+
+export const Header = () => {
+    return (
+        <div className='header'>
+            <header>Happy 💌 Thought</header>
+        </div>
+    )
+}
